Do not check for govuk:content-id meta tag
Not all pages have this tag, for example /contact and
/government/how-government-works.  These are both "special_route"
documents, where the corresponding content item is empty (beyond a
dummy content ID and the route).

Making apps talk to the content store to find out the (unimportant!)
content ID of a special_route page feels like a needlessly complex
change for very little benefit.