Always exit after n failed retries for Whitehall
This provides better feedback from the system since a quiet failure earlier in the chain can cause a loud failure later on and skew the results.  This causes the chain to stop immediately if that step's retry limit is reached.

I spent some time attempting to extend gatling to include both the retries and failure command in one method, but was unsuccessful due to ignorance of scala and/or the gatling classes not being designed for extension.