Add a scenario for Coronavirus local restrictions
This is a rushed together scenario for testing the
/find-coronavirus-local-restrictions endpoint. I don't remotely know
what I'm doing with Gatling so I put this together by copy and pasting
an existing test and changing as little as possible.

I've randomly generated a list of 100,000 postcodes for testing. I'm
kinda hoping it doesn't choke on the + in the URL.