Basic school closure test
This is a test for load testing the quantity of POST requests can be
handled to a particular local transaction, in this case
/school-closure-check. The user process to access these involves accessing
a form, submitting it and then being redirected to a local authority page.

We expect that the form itself will be mostly cached at the CDN and that
there are a limited number of local authority pages that will also
mostly likely be cached, therefore the area to really test is the post
requests that always make it to origin.

This only tests a single postcode as there isn't any caching of
postcodes in the app so testing one postcode repeatedly should have the
same affect as testing a variety of them.